Moved manual bonus products processing from the cart page to the bonus products specific hook.
