+++ packages/template-chakra-storefront/src/pages/cart/index.jsx
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import React, {useEffect, useRef, useCallback} from 'react'
+import React from 'react'
 import {Box, Stack, Grid, GridItem, Container, useDisclosure} from '@chakra-ui/react'
 import {useCurrentBasket, useCurrentCustomer} from '../../hooks/'
 
@@ -37,104 +37,22 @@ const Cart = () => {
     // Product data and processing
     const {isProductsPending, productsByItemId} = useCartProducts(basket)
 
-    // Initialize manual bonus products hook
+    // Initialize manual bonus products hook - handles all processing automatically
+    useManualBonusProducts(basket, isPending, isRegistered)
+
+    // Cart operations
     const {
-        manualBonusProductCollections,
-        createManualBonusProductCollections,
-        detectNewlyAddedBonusProducts,
-        analyzeQualifyingProductChanges,
-        clearAllManualBonusProductCollections
-    } = useManualBonusProducts()
-
-    // Store previous basket state for comparison
-    const prevBasketRef = useRef(null)
-
-    // Effect to detect and track bonus products when basket changes
-    useEffect(() => {
-        if (!basket || isPending) return
-
-        const previousBasket = prevBasketRef.current
-        if (!previousBasket) {
-            prevBasketRef.current = basket
-            return
-        }
-
-        // Analyze changes in qualifying products
-        const qualifyingProductChanges = analyzeQualifyingProductChanges(
-            previousBasket,
-            basket,
-            [] // addedProductIds - this would come from add-to-cart operations
-        )
-
-        if (qualifyingProductChanges.length > 0) {
-            // Detect newly added bonus products
-            const detectionResult = detectNewlyAddedBonusProducts(
-                previousBasket,
-                basket,
-                qualifyingProductChanges
-            )
-
-            // Create/update manual bonus product collections
-            if (detectionResult.qualifyingProductToBonusProducts) {
-                createManualBonusProductCollections(
-                    detectionResult.qualifyingProductToBonusProducts
-                )
-            }
-
-            console.log('Manual Bonus Products Updated:', {
-                qualifyingProductChanges: detectionResult.qualifyingProductChanges,
-                newBonusProducts: detectionResult.newBonusProducts,
-                collections: manualBonusProductCollections
-            })
-        }
-
-        // Update previous basket reference
-        prevBasketRef.current = basket
-    }, [
-        basket,
-        isPending,
-        analyzeQualifyingProductChanges,
-        detectNewlyAddedBonusProducts,
-        createManualBonusProductCollections,
-        manualBonusProductCollections
-    ])
-
-    // Clear collections when basket is empty or customer logs out
-    useEffect(() => {
-        if (!basket?.productItems?.length || !isRegistered) {
-            clearAllManualBonusProductCollections()
-        }
-    }, [basket?.productItems?.length, isRegistered, clearAllManualBonusProductCollections])
-
-    // Cart operations with bonus product tracking integration
-    const cartOperations = useCartOperations(basket, productsByItemId, showError)
-
-    // Enhanced quantity change handler that tracks qualifying products
-    const handleChangeItemQuantityWithTracking = useCallback(
-        async (product, value) => {
-            // Call original handler
-            const result = await cartOperations.handleChangeItemQuantity(product, value)
-
-            // The basket will be updated via the useCurrentBasket hook
-            // The useEffect above will automatically detect changes and update bonus collections
-
-            return result
-        },
-        [cartOperations.handleChangeItemQuantity]
-    )
-
-    // Enhanced remove item handler
-    const handleRemoveItemWithTracking = useCallback(
-        async (product) => {
-            // Call original handler
-            const result = await cartOperations.handleRemoveItem(product)
-
-            // The useEffect above will automatically detect changes and update bonus collections
-
-            return result
-        },
-        [cartOperations.handleRemoveItem]
-    )
+        selectedItem,
+        setSelectedItem,
+        localQuantity,
+        isCartItemLoading,
+        setCartItemLoading,
+        handleUpdateCart,
+        handleUpdateBundle,
+        handleUnavailableProducts,
+        handleChangeItemQuantity,
+        handleRemoveItem
+    } = useCartOperations(basket, productsByItemId, showError)
 
     // Wishlist operations
     const {addToWishlist} = useWishList()
@@ -143,8 +61,8 @@ const Cart = () => {
     // Gift items
     const {localIsGiftItems, handleIsAGiftChange} = useCartGiftItems(
         basket,
-        cartOperations.setCartItemLoading,
-        cartOperations.setSelectedItem,
+        setCartItemLoading,
+        setSelectedItem,
         showError
     )
 
@@ -156,14 +74,14 @@ const Cart = () => {
 
     // Handle edit click
     const handleEditClick = (product) => {
-        cartOperations.setSelectedItem(product)
+        setSelectedItem(product)
         onOpen()
     }
 
     // Handle modal close
     const handleModalClose = () => {
         onClose()
-        cartOperations.setSelectedItem(undefined)
+        setSelectedItem(undefined)
     }
 
     /********* Rendering UI **********/
@@ -195,16 +113,16 @@ const Cart = () => {
                                 <CartProductList
                                     basket={basket}
                                     productsByItemId={productsByItemId}
-                                    localQuantity={cartOperations.localQuantity}
+                                    localQuantity={localQuantity}
                                     localIsGiftItems={localIsGiftItems}
                                     isProductsPending={isProductsPending}
-                                    isCartItemLoading={cartOperations.isCartItemLoading}
-                                    selectedItem={cartOperations.selectedItem}
-                                    handleChangeItemQuantity={handleChangeItemQuantityWithTracking}
+                                    isCartItemLoading={isCartItemLoading}
+                                    selectedItem={selectedItem}
+                                    handleChangeItemQuantity={handleChangeItemQuantity}
                                     handleIsAGiftChange={handleIsAGiftChange}
                                     handleAddToWishlist={handleAddToWishlist}
                                     handleEditClick={handleEditClick}
-                                    handleRemoveItem={handleRemoveItemWithTracking}
+                                    handleRemoveItem={handleRemoveItem}
                                 />
                             </GridItem>
                             <GridItem>
@@ -226,12 +144,12 @@ const Cart = () => {
                 isOpen={isOpen}
                 onOpen={onOpen}
                 onClose={handleModalClose}
-                selectedItem={cartOperations.selectedItem}
-                handleUpdateCart={cartOperations.handleUpdateCart}
-                handleUpdateBundle={cartOperations.handleUpdateBundle}
-                handleRemoveItem={handleRemoveItemWithTracking}
+                selectedItem={selectedItem}
+                handleUpdateCart={handleUpdateCart}
+                handleUpdateBundle={handleUpdateBundle}
+                handleRemoveItem={handleRemoveItem}
                 basket={basket}
-                handleUnavailableProducts={cartOperations.handleUnavailableProducts}
+                handleUnavailableProducts={handleUnavailableProducts}
             />
         </Box>
     )
